apply requested style changes
