+++ quinn-udp/src/runtime/tokio_runtime.rs
@@ -1,11 +1,17 @@
-use super::{AsyncWrappedUdpSocket, Runtime};
 use std::io;
+use std::net::{SocketAddr, UdpSocket as StdUdpSocket};
 use std::task::{Context, Poll};
+
+use super::{AsyncUdpSocket, Runtime};
+
 #[cfg(unix)]
 use tokio::io::unix::AsyncFd;
 
+#[derive(Debug, Clone)]
+pub struct TokioRuntime;
+
 #[cfg(unix)]
-impl AsyncWrappedUdpSocket for AsyncFd<std::net::UdpSocket> {
+impl AsyncUdpSocket for AsyncFd<StdUdpSocket> {
     fn poll_read_ready(&self, cx: &mut Context) -> Poll<io::Result<()>> {
         AsyncFd::poll_read_ready(self, cx).map(|x| x.map(|_| ()))
     }
@@ -30,25 +36,25 @@ impl AsyncWrappedUdpSocket for AsyncFd<std::net::UdpSocket> {
         }
     }
 
-    fn try_recv_from(&self, buf: &mut [u8]) -> io::Result<(usize, std::net::SocketAddr)> {
+    fn recv_from(&self, buf: &mut [u8]) -> io::Result<(usize, SocketAddr)> {
         self.get_ref().recv_from(buf)
     }
 
-    fn try_send_to(&self, buf: &[u8], target: std::net::SocketAddr) -> io::Result<usize> {
+    fn send_to(&self, buf: &[u8], target: SocketAddr) -> io::Result<usize> {
         self.get_ref().send_to(buf, target)
     }
 
-    fn local_addr(&self) -> io::Result<std::net::SocketAddr> {
+    fn local_addr(&self) -> io::Result<SocketAddr> {
         self.get_ref().local_addr()
     }
 
-    fn get_ref(&self) -> &std::net::UdpSocket {
+    fn get_ref(&self) -> &StdUdpSocket {
         AsyncFd::get_ref(self)
     }
 }
 
 #[cfg(not(unix))]
-impl AsyncWrappedUdpSocket for tokio::net::UdpSocket {
+impl AsyncUdpSocket for tokio::net::UdpSocket {
     fn poll_read_ready(&self, cx: &mut Context) -> Poll<io::Result<()>> {
         tokio::net::UdpSocket::poll_recv_ready(self, cx)
     }
@@ -62,30 +68,24 @@ impl AsyncWrappedUdpSocket for tokio::net::UdpSocket {
     }
 
     fn clear_write_ready(&self, _cx: &mut Context) {
-        // not necessary because tokio::net::UdpSocket::try_send_from already uses try_io
+        // not necessary because tokio::net::UdpSocket::try_send_to already uses try_io
     }
 
-    fn try_recv_from(&self, buf: &mut [u8]) -> io::Result<(usize, std::net::SocketAddr)> {
+    fn recv_from(&self, buf: &mut [u8]) -> io::Result<(usize, SocketAddr)> {
         tokio::net::UdpSocket::try_recv_from(self, buf)
     }
 
-    fn try_send_to(&self, buf: &[u8], target: std::net::SocketAddr) -> io::Result<usize> {
+    fn send_to(&self, buf: &[u8], target: SocketAddr) -> io::Result<usize> {
         tokio::net::UdpSocket::try_send_to(self, buf, target)
     }
 
-    fn local_addr(&self) -> io::Result<std::net::SocketAddr> {
+    fn local_addr(&self) -> io::Result<SocketAddr> {
         tokio::net::UdpSocket::local_addr(self)
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct TokioRuntime;
-
 impl Runtime for TokioRuntime {
-    fn wrap_udp_socket(
-        &self,
-        t: std::net::UdpSocket,
-    ) -> io::Result<Box<dyn AsyncWrappedUdpSocket>> {
+    fn wrap_udp_socket(&self, t: StdUdpSocket) -> io::Result<Box<dyn AsyncUdpSocket>> {
         t.set_nonblocking(true)?;
         #[cfg(unix)]
         {
