--- quinn/src/runtime/async_std_runtime.rs
@@ -1,26 +1,0 @@
-use super::{AsyncTimer, Runtime};
-use async_io::Timer;
-use std::future::Future;
-use std::pin::Pin;
-use std::task::{Context, Poll};
-use std::time::Instant;
-
-impl AsyncTimer for Timer {
-    fn reset(mut self: Pin<&mut Self>, t: Instant) {
-        self.set_at(t)
-    }
-    fn poll(self: Pin<&mut Self>, cx: &mut Context) -> Poll<()> {
-        Future::poll(self, cx).map(|_| ())
-    }
-}
-
-pub use udp::runtime::AsyncStdRuntime;
-
-impl Runtime for AsyncStdRuntime {
-    fn new_timer(&self, t: Instant) -> Pin<Box<dyn AsyncTimer>> {
-        Box::pin(Timer::at(t))
-    }
-    fn spawn(&self, future: Pin<Box<dyn Future<Output = ()> + Send>>) {
-        async_std::task::spawn(future);
-    }
-}
