--- quinn/src/runtime/tokio_runtime.rs
@@ -1,27 +1,0 @@
-use super::{AsyncTimer, Runtime};
-use std::future::Future;
-use std::pin::Pin;
-use std::task::{Context, Poll};
-use std::time::Instant;
-use tokio::time::{sleep_until, Sleep};
-
-impl AsyncTimer for Sleep {
-    fn reset(self: Pin<&mut Self>, t: Instant) {
-        Sleep::reset(self, t.into())
-    }
-    fn poll(self: Pin<&mut Self>, cx: &mut Context) -> Poll<()> {
-        Future::poll(self, cx)
-    }
-}
-
-pub use udp::runtime::TokioRuntime;
-
-impl Runtime for TokioRuntime {
-    fn new_timer(&self, t: Instant) -> Pin<Box<dyn AsyncTimer>> {
-        Box::pin(sleep_until(t.into()))
-    }
-
-    fn spawn(&self, future: Pin<Box<dyn Future<Output = ()> + Send>>) {
-        tokio::spawn(future);
-    }
-}
